Created skeleton for web project
The website now contains a general skeleton off of which to devlop the
project. The navigation bar contains MenuList, Friends, and Profile
pages which will be further devloped. The backend code also contains
route and model examples, which can help for implementing further
backend structures.

The following is the general structure of the git directory:

/mynu - This is the name of the frontend directory. It contains all
documents relevant to frontend development

/mynu/src - This directory contains the source code for the frontend

/mynu/src/components - This directory contains the react components

/backend - This directory It contains all  documents relevant to
backend development

/backend/models - This directory contains the models for the data
structures that are stored in the DB

/backend/routes - This directory contains the routes for API requests

Building:
* To run the frontend run the command 'npm start' from /mynu
* To run the backend run the command 'nodemon server' from /backend

Important notes:
* You will need to run the command 'npm install' in the /mynu and
/backend directories in order to ensure that you have the needed
packages
* In order to run the backend, you will need to copy the ATLAS_URI which
can be accessed through the Mongo DB web portal.