Added Rating Feature and fixed npm packaging
Added a dislike and like feature. The feature is currently available to
all users, with no the limit to the number of likes and dislikes only
per visit to page.
TODO: Limit the feature to when the user is logged in and limit the
quantity to one per user.

The MenuItem page currently displays the information, but not in a
particularly aesthetically pleasing way.
TODO: Update the aesthetics

Fixed npm packaging on branch that was a result of mistaken commits
that pushed a node_modules folder and unneccsary packages. Also,
changed a .env file that was mistakenly pushed to main.